+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [3.1]
+### - 2024-04-21
+- Draggable Plugins New Add
+
 ### - 2024-04-17
 - RecyclerView 업데이트, 테이블뷰 테스트 및 스크롤 테스트
 - CacheMomory 버그 패치 및 예제 업데이트